+++ comics/comics/brinkerhoff.py
@@ -6,17 +6,9 @@ class Meta(MetaBase):
     language = 'en'
     url = 'http://www.brinkcomic.com/'
     start_date = '2006-01-02'
+    end_date = '2009-12-30'
     rights = 'Gabe Strine'
 
 class Crawler(CrawlerBase):
-    history_capable_date = '2006-01-02'
-    time_zone = -5
-
     def crawl(self, pub_date):
-        url = 'http://www.brinkcomic.com/comics/%s.gif' % (
-            pub_date.strftime('%Y%m%d'),)
-        headers = {
-            'Referer': 'http://www.brinkcomic.com/d/%s/' % (
-                pub_date.strftime('%Y%m%d'),)
-        }
-        return CrawlerImage(url, headers=headers)
+        pass # Comic no longer published